+++ apps/frontend/src/libs/components/components.ts
@@ -1,6 +1,7 @@
 export { App } from "./app/app.js";
 export { Button } from "./button/button.js";
 export { Header } from "./header/header.js";
+export { Icon } from "./icon/icon.js";
 export { Input } from "./input/input.js";
 export { Link } from "./link/link.js";
 export { Loader } from "./loader/loader.js";
@@ -9,5 +10,6 @@ export { ProgressBar, type Step } from "./proggress-bar/progress-bar.js";
 export { ProtectedRoute } from "./protected-route/protected-route.js";
 export { QuizQuestion } from "./quiz-question/quiz-question.js";
 export { RouterProvider } from "./router-provider/router-provider.js";
+export { Sidebar } from "./sidebar/sidebar.js";
 export { Provider as StoreProvider } from "react-redux";
 export { Navigate, Outlet as RouterOutlet } from "react-router-dom";