+++ apps/frontend/src/libs/components/icon/libs/maps/icon-name-to-svg.map.ts
@@ -1,0 +1,17 @@
+import CrossedEye from "~/assets/img/crossed-eye.svg?react";
+import Eye from "~/assets/img/eye.svg?react";
+import WheelOfBalanceActive from "~/assets/img/wheel-of-balance-active.svg?react";
+import WheelOfBalanceInactive from "~/assets/img/wheel-of-balance-inactive.svg?react";
+import { type IconName } from "~/libs/types/types.js";
+
+const iconNameToSvg: Record<
+	IconName,
+	React.FC<React.SVGProps<SVGSVGElement>>
+> = {
+	crossedEye: CrossedEye,
+	eye: Eye,
+	wheelActive: WheelOfBalanceActive,
+	wheelInactive: WheelOfBalanceInactive,
+};
+
+export { iconNameToSvg };