+++ apps/frontend/src/libs/constants/sidebar-items.constant.ts
@@ -1,0 +1,15 @@
+import { AppRoute } from "~/libs/enums/enums.js";
+import { type SidebarItem } from "~/libs/types/types.js";
+
+const SIDEBAR_ITEMS: SidebarItem[] = [
+	{
+		href: AppRoute.ROOT,
+		icon: {
+			active: "wheelActive",
+			inactive: "wheelInactive",
+		},
+		label: "My Wheel",
+	},
+];
+
+export { SIDEBAR_ITEMS };