+++ apps/frontend/src/libs/types/types.ts
@@ -1,4 +1,6 @@
 export { type AsyncThunkConfig } from "./async-thunk-config.type.js";
+export { type IconName } from "./icon-name.type.js";
+export { type SidebarItem } from "./sidebar-item.type.js";
 export {
 	type Control,
 	type FieldErrors,