+++ apps/mobile/android/app/src/main/java/com/bebalance/MainActivity.kt
@@ -1,11 +1,31 @@
 package com.bebalance
 
+import android.os.Bundle
+import android.view.View
+import android.view.WindowManager
 import com.facebook.react.ReactActivity
 import com.facebook.react.ReactActivityDelegate
 import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint.fabricEnabled
 import com.facebook.react.defaults.DefaultReactActivityDelegate
 
 class MainActivity : ReactActivity() {
+
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(savedInstanceState)
+
+    // Make the navigation bar transparent
+    window.apply {
+      clearFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_NAVIGATION)
+      addFlags(WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS)
+      navigationBarColor = android.graphics.Color.TRANSPARENT
+      decorView.systemUiVisibility = (
+        View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
+        or View.SYSTEM_UI_FLAG_LAYOUT_STABLE
+        or View.SYSTEM_UI_FLAG_LIGHT_NAVIGATION_BAR
+        or View.SYSTEM_UI_FLAG_LIGHT_STATUS_BAR
+      )
+    }
+  }
 
   /**
    * Returns the name of the main component registered from JavaScript. This is used to schedule