+++ apps/mobile/src/libs/types/navigation/root-navigation-parameter-list.type.ts
@@ -1,6 +1,7 @@
 import { type RootScreenName } from "~/libs/enums/enums";
 
 type RootNavigationParameterList = {
+	[RootScreenName.QUIZ_ENTRY]: undefined;
 	[RootScreenName.SIGN_IN]: undefined;
 	[RootScreenName.SIGN_UP]: undefined;
 };