+++ apps/mobile/src/navigations/root/root.tsx
@@ -5,8 +5,10 @@ import {
 import React from "react";
 
 import { RootScreenName } from "~/libs/enums/enums";
+import { useAppSelector } from "~/libs/hooks/hooks";
 import { type RootNavigationParameterList } from "~/libs/types/types";
 import { Auth } from "~/screens/auth/auth";
+import { QuizEntry } from "~/screens/quiz-entry/quiz-entry";
 
 const NativeStack = createNativeStackNavigator<RootNavigationParameterList>();
 
@@ -15,10 +17,22 @@ const screenOptions: NativeStackNavigationOptions = {
 };
 
 const Root: React.FC = () => {
+	const user = useAppSelector((state) => state.auth.user);
+	const hasUser = Boolean(user);
+
 	return (
 		<NativeStack.Navigator screenOptions={screenOptions}>
-			<NativeStack.Screen component={Auth} name={RootScreenName.SIGN_IN} />
-			<NativeStack.Screen component={Auth} name={RootScreenName.SIGN_UP} />
+			{hasUser ? (
+				<NativeStack.Screen
+					component={QuizEntry}
+					name={RootScreenName.QUIZ_ENTRY}
+				/>
+			) : (
+				<NativeStack.Group>
+					<NativeStack.Screen component={Auth} name={RootScreenName.SIGN_IN} />
+					<NativeStack.Screen component={Auth} name={RootScreenName.SIGN_UP} />
+				</NativeStack.Group>
+			)}
 		</NativeStack.Navigator>
 	);
 };