+++ packages/shared/src/index.ts
@@ -29,7 +29,10 @@ export {
 	type ValidationSchema,
 	type ValueOf,
 } from "./libs/types/types.js";
-export { AuthApiPath } from "./modules/auth/auth.js";
+export {
+	AuthApiPath,
+	ConfirmPasswordCustomValidation,
+} from "./modules/auth/auth.js";
 export {
 	type SurveyAnswerDto,
 	SurveyApiPath,
@@ -37,12 +40,12 @@ export {
 } from "./modules/surveys/surveys.js";
 export {
 	type UserDto,
-	type UserGetAllItemResponseDto,
 	type UserGetAllResponseDto,
 	UsersApiPath,
 	type UserSignInRequestDto,
 	type UserSignInResponseDto,
 	userSignInValidationSchema,
+	type UserSignUpFormDto,
 	type UserSignUpRequestDto,
 	type UserSignUpResponseDto,
 	userSignUpValidationSchema,