+++ packages/shared/src/modules/users/libs/types/user-sign-up-form-dto.type.ts
@@ -1,0 +1,5 @@
+import { type UserSignUpRequestDto } from "./types.js";
+
+type UserSignUpFormDto = { confirmPassword: string } & UserSignUpRequestDto;
+
+export { type UserSignUpFormDto };